Feat: chatgpt api 구현 - chatgpt api 질문 수에 관계 없이 한 번에 분석되도록 구현함 S12P11B103-246

+++ BackEnd/mindon/src/main/java/com/ssafy/mindon/meeting/domain/UserReview.java
@@ -24,7 +24,7 @@ public class UserReview {
     private Integer meetingId;
 
     @Column(name = "meeting_week", nullable = false)
-    private Integer meetingWeek;
+    private Byte meetingWeek;
 
     @Column(name = "summation", columnDefinition = "TEXT", nullable = false)
     private String summation;
